Split story_callbacks mod into story submodules
