+++ lib/src/story/progress.rs
@@ -10,8 +10,7 @@ use crate::{
     object::RTObject,
     pointer::{self, Pointer},
     push_pop::PushPopType,
-    story::{OutputStateChange, Story},
-    story_callbacks::ErrorType,
+    story::{errors::ErrorType, OutputStateChange, Story},
     story_error::StoryError,
     story_state::StoryState,
     tag::Tag,
